refactor(api): Move api-hooks.ts from api to api/impl
